+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.0
+- Refactored with Template Method pattern,
+- allowing customization of record and format methods.
+
 ## 0.0.3
 
 - supports all six platforms, with file logging separated for non-web platforms.